Internal - Reorganization of the database script folder
I've reorganized so that there is only one script to run. Folders will be added to create new individual update scripts but they will be rolled into the the cumulative one. To do this each individual script should have the necessary checks to ADD/UPDATE/DROP as necessary for the script to run without user interaction and without errors. For example when creating a new table, the table should be checked for existance before creating otherwise an error will be generated to the user.