Merged revision(s) 2037-2179 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk:
Merge of v13.0 trunk changes into v13.1. It includes the following updates:

Internal - Reorganization of the database script folder

I've reorganized so that there is only one script to run. Folders will be added to create new individual update scripts but they will be rolled into the the cumulative one. To do this each individual script should have the necessary checks to ADD/UPDATE/DROP as necessary for the script to run without user interaction and without errors. For example when creating a new table, the table should be checked for existance before creating otherwise an error will be generated to the user.
........
Internal - Reverting accidental deletion of the v2.9 to v3.0 script.
........
Bug #946 - Cannot view upload files

This does not seem to be an issue with the web pages but an issue with configuration. I did take the time to bulletproof the code for when there is no systems defined as the home.asp page will generate an error when there are no systems.
........
Update version file to v3.0.4.2 10/28/2013 for Release.
........