Bug #965 - Documents not showing on Customer page
Reported by First Federal Savings Bank. The issue was tracked down to the evaluation of the SQL WHERE clause for filtering out documents. The issue was the groupActivation string was doing a LIKE comparison but was failing. The condition has been changed to an Equality condition which does work. May be an issue with 2012 as we've not been able to recreate with 2008.