Ticket 931 - Scan New Document Option "Disappears" When Using Edit Document Tab (2013.0.0.18)
The initial fix did worked but it was found an error in the same process as when you select "Loan", "Deposit" or "Trusts".  The initial error was only documented for "Credit"

To fix this I have had to modify the following programs

a) documentactions.asp - making sure that the code is accurate because I used the page for testing
b) include/uploadnavpanel.inc    (lines 72 to 76)
c) include/cust_accountdocbody.inc  (lines 269 to 283)
d) include/custloandoc.asp   (lines 270 to 283)

I made sure that the "customerId" and "loadId" were passed 100% of the time.

James