Ticket 962 - Currency validation does not work on larger currency values.
The error has been caused not because of the code but the actual length of the field in the "loan" table.

The database change that is needed on the "loan" table for the "loanAmount" column.

The current data type is    "nvarchar(12)" and it needs to be   "nvarchar(20)".

I have made the change to the web page  "loanmaintdetails.asp"  so that the max allowed will be 20.

James