Ticket 921 - AccuSystems RSS news feed is broken.
I added an entry into the "accuSystemsProperties" table with the "PropertyKey" value of "acculoan.RSSUrl" so that the URL is stored there.

I updated the program "dashpannelrss.asp" to read the URL from the session value "acculoan.RSSUrl".

I wrote a script that I have attached to this Ontime Ticket to update the database table "accuSystemsProperties" to have the new "PropertyKey" value

I have passed this ticket on to Andrew so that he can manage the database update.

James