Backlog #226 - Upload does not work with Aquaforest
The upload link still had the code to send the image files directly to the browser. I've modified the page to pass the necessary information to the viewer-selector.asp page. Also the viewer-selector.asp page had to be modified to handle files coming from the scanned images folder versus the upload folder.