Ticket 415 - Admin User can edit/create/maintenance regions/banks/branches (test 2 MultiBank)
I modified the program "bank_list.asp" to check the function "IsAdministrator()" to determine if the user is an Administrator.  Lines 106 - 114

I have added the condition to only allow the administrator to be able to edit a bank only if they have been granted access to the bank via the "Bank/Branch Security" option of the "User Maintenance".  This code  reads the "userBankSecurity" table to verify if the user has bank access.

James