Bug #842 - Dashpanels not saving settings consistently
This was due to code that attempted to allow multiple pages of panels beyond the first 6. I've corrected the code so that it will only use 6 and a script to clean the bad data from the userDashboard table. I've added some convenient constants that if we wish to allow more pages in the future it should be possible, but adding the extra UI is beyond the scope of this ticket.