Bug #988 - Error opening AccuAccount from AccuDoc Merged from v13.0.23 maintenance branch
The error is in AccuAccount externalportal.asp page (duplicated variable declaration) and is fixed there. The change to AccuDoc is to deal with opening AccuAccount but not as a frame of AccuDoc. I've added javascript to ensure when AccuAccount opens, it will open AccuAccount properly and not within AccuDoc's main frame.