Bug #999 - Aquaforest not called when view document history TIFF files
This was broken due to the View Tracking feature added. The filename was no longer being passed and so the logic assumed that all files were not TIFF files. I've added the extra logic to extract the filename that is generated from the documentId so the check occurs properly now.