Merged revision(s) 8-2339, 2595-2616 from AccuSystems/Products/Development/AccuLoan/Apps/Web/branches/v13.0.23: Bug #1027 - Converting Application into existing Booked Loan not copying files over
This was due to the document redesign changes in that the assumed logic expected documentTab records to exist to move the documents to. The fix was into insert the necessary records before the move and correct the logic that copied the files to the new location. Also I've modified the wizard that allows the selection of the booked loan to not make available Loans of a different type as the Application as the files will never be copied over and to avoid this situation it's safer to not allow it.