Bug #1000 - Document definition maintenance screen jumbled
This was due to conflicting third party code that relies heavily on CSS and Javascript. I've modified the maintenance page for the Visibility Rules to not open in the GreyBox window. The GreyBox, when closed would re-render the area underneath, but much of the content underneath is hidden due the DHTMLX Tabs being used. GreyBox could tell what is visible and what isn't. The solution is to have the Visibilty Rules be open within the normal browser window for now as there's no quick solution to resolve GreyBox and DHTMLX Tab conflicts.