Bug #1026 - Issue with QC rejection on scan history Merged from v13.0.24
This issue is due to documentHistory documentId fields not being NULLed out when the document record does not exist anymore. This usually happens over time, in this case a document is rejected and moved to a new tab. The old history record keeps the documentId field when it should be nulled out because the document record does not exist anymore for that record. I've added a cleanup script to NULL any documentId field if it does not exist in the document table.