OnTime ticket #1055. Updated clsDocument for alternate barcodes. (originally changed by amiller but in the incorrect project. I've manually merged the changes into this project)
The issue was that the regular expression for alternate barcodes are used to find a match and then the the barcode value is used to lookup the document code.  This works fine if the document code matches the entire barcode value.  If the document code is a subset of the barcode, accuscan still tries to used the entire barcode value to lookup the document code.  Added code to the getInfoFromDb function so that if a code isn't found when using the barcode value, it will also attempt to lookup the code based on the regex match value.