Merged revision(s) 8-2339, 2341-2615, 2821-2825 from AccuSystems/Products/Development/AccuLoan/Apps/Web/branches/v13.0.24:
This merge is by James for this ticket.

Ticket 1065 - Issue with the way the ‘Copy Loan Documents’ feature reports in the View Scan History

I added "l.IsCollateralYN" to the values that were to be retreived from the query "srcDocumentQuery" on line 83.

This variable is being used in the code the builds the text line which starts on line 313 "insertCmd".

James