Bug #1052 - Quick move function does not move document to customers
This was an issue in the Quick Move page allowing the user to select the target document to move to. The list was not constrained by documents for the given Customer's bank. Hence it was possible to select a document that belonged to a different bank. This was noticeable due to the duplicate Tabs displayed. Once the list of document tabs were constrained by the Customer's bank, the document moved correctly.