Bug #854 - Flex Field activation status confusing
I've modified the Flex Field Definition maintenance page hopefully resolve the confusing about the Activation status flag. The listing and the maintenance form now indicate the Default Activation Status of the Flex Field for the given Customer/Account/Collateral. This does not override Customers/Accounts/Collaterals if altered by a user. It just indicates the Default status if it has not been set yet. The check box still allows user to set all instance Flex Fields with the default value if they need to.