Ticket 978 - Unable to edit Account information on Edit page after adding a customer manually
I have included the following line
    <!-- #include file="include/custmaint_accounts.asp" -->
on line 1479.

I removed the code for this section of the customermaint.asp progam and put it in the above include to control the modification and enhanchments.

I created two triggers 
a) enableActionDisplayData
b) enableActionDisplayOnly
So, that the processing will know if there is any data to display.

Then the enableActionDisplayData would control an outer loop of processing.

James