Ticket 978 - Unable to edit Account information on Edit page after adding a customer manually
I modified this program 
    <!-- #include file="include/custmaint_accounts.asp" -->

I made the change to allow the showing of data if it was present.  When I first made the change I fixed the ticket to show the tabs correct when there was a new account but when there was an exisiting account the data was not displaying.  Now it does correctly.

James