Bug #1115 - Audit displaying N/A documents with no files
This was a subtle bug in the SQL that generates the document listing for the Credit/Account/Collateral document sections. The issue occurs when no document records exist because they are defaulting, but the group for a particular Customer/Account/Collateral Document Group is overridden to be inactive. The activation status was being joined on the documents which may or may not exist. When they didn't exist the overridden activation status was not used. I've modified the SQL to ensure the Document Activations are joined properly with actual Customer/Account/Collateral records so the data is reflected accurately.