Bug 1123 - Woodforest- EV will not set new Customers/Loans to have Exceptions
The reason for the ExceptionValidator not working is due to the LQAS custom exceptions imported having DateTime values for the Reminder Date when added to the Grace Period overflow the maximum date value the program can manage. The fix was to catch the error and use the maximum data time value allowed.