Bug#1097 - CBC View Scan history issue for view non-employee file documents
This bug was due to the search results no taking into account the Document Tab's HideEmployeeFile flag. The links are no available for non-hidden employee document tabs. Also a new icon to show that the document is there but grayed out when not accessible by the employee.

Update: Added logic to allow Admin Users access to employee files based on the Account Class the document belongs to