Ticket 644 - User is able to view docs from banks not assigned to them in scan history (Multi Banks Test 2)
Updated the program "sh_searchresults.asp" to add "include/getUserBankSecurityArray.inc" to be able to use the array and the function "HasBankAccess" to determine if the user has access to the bank where the data is to be displayed.

I also added the if condition "if HasBankAccess(detailRS("bankId"))" within the loop to be able to control the access.  Line 296

James