Bug_787 - Search Panel - Account Number - Test 3 Site
Modified the program  "dhtmlxSearchValues.asp"  starting at line 90 because when the "Account Number" is being searched the system reads the "loan" table.  Based on MultiBank there is not a value "bankId" in the loan table.  As a result the query would fail.

I make a specific condition to check if the "loan" table was selected and if so perform a different query.

James