Merged revision(s) 3234-3254 from AccuSystems/Products/Development/AccuLoan/Apps/Web/branches/v13.0.27/WebSite: Bug_787 - Search Panel - Account Number - Test 3 Site
Modified the program  "dhtmlxSearchValues.asp"  to update the SQL statement to take into account MultiBank access.  LInes 91 to 99.

James
........
Started to add a few general styles that can help the site by reducing inline styling.
........
Disabled buffering to prevent the buffer from being exceeded.
........
Started to add a few general styles that can help the site by reducing inline styling. This is v14
........
Added a javascript validation that requests data if none is entered.
........
Created a new include file that contain all the lines of code necessary for dhtmlcombo box
........
Ticket 1118 : Added a javascript validation that requests data if none is entered.
........
Adjusted to tidy and streamline code
........
Ticket 1117 : Added a javascript validation that requests data if none is entered. Updated Code to avoid HTML coding errors!
........
Removing declaration of extendedAccountClassCode in uploadnavpanel.inc because its defined in it's including parent page for security lookups.
........
Added declaration of extendedAcccountClassCode for fix in the "view full upload" link in dashpanel.
........
Tickets: 1117, 1118, and 1119

Added Required fields (*) Asterisk to give the user visual references as to which fields are required.
........
Ticket 1159: Upload View Full Page Error displayed when changing folders

Minor syntax error corrected. Response:Write was changed to Response.Write
........
Ticket 1158: Notice Letters Next Notice Date Generates Error

To prevent the error from showing when it is todays date, I subtracted 1 day from date in milliseconds so it is comparing todays date with yesterday's date during the validation.
........
Reverting this page to Rev 3156

This is to merge the changes from v13 to v14 properly.
........