Bug #793, #1142 - Merging Tiff to PDF causes checksum error Merged from v13.0.28
This error was due to how TIF files are managed when there is only one page. The GDPicture did not return a valid page count when there is only one page because Tifs treat a "single" page image differently than a multi-page image. Extra logic has been added the AccuSystemsCommonLib.dll to ensure the check for single page occurs and appends the image appropriately.

Update: Needed to append random information at the end of the file URL to force the browser not to use the cached image of the file.