Backlog #1195 - Upload Navigation Enhancement for CBC
Initial Development testing has been completed on this new feature and is being checked in for QC Testing. The detailed design document can be viewed in OnTime for the given ticket number to see detailed changes to the various web files.

Update: Added database schema change files to SVN for new enhancement.