Bug_859 - User Perm - User with Read/Edit record permissions allowed to activate/inactivate groups/change expiration dates etc
I worked to find out why the icon "Edit Doc Tab" was not showing for the Credit side of the customer page and found that it was because of bad data.  There is a condition that if the logged in user does not have access to the branch that the account is in the icon "Edit Doc Tab" will not be displayed.  

In the Customer table the column "customerBranchId" column value had a value for a branch that did not exist in the branch table.

I have informed Andy about it as well so that he can create a dbTest to validate that this does not exist.

I have also modified the program "include/cust_collateraldocbody.inc"  (line 230) to add ".allowDocRead" so that the "Edit Doc Tab" appears for Collateral.

James