Bug #1228 - Search redirects to wrong account number when customer has similar accounts. Merged from v13.0.29
Added code back in to only redirect if there is a single customer and single loan in the search results. This avoids the issue when looking up an Account Numbers that are similar for the same customer. For example looking up Acccount Number 1234 when a Customer has another account like 101234. The Search results sometimes will pick the incorrect account, because of wildcarding.