Ticket 1215 | Lender Field in Applications Not showing New Lenders Created Version 14
When creating a new user, the "Inactive" dropmenu was disabled and thus did not send a "N" value to the database creating a user without inactive status (NULL).  This inactive status is used for the lender list so it doesn't show that user in this list of lenders. I removed "disabled" to ensure the database adds N as the inactive status and this resolved the problem.