Bug_1135 - CBC - Sort on View Account Listing
I modified the code for program "accoutlist.asp" to do the following:  
(a) Zero fill the day and month for all single numbers
(b) Change the display layout to be YYYY/MM/DD
so that the sort work 100% of the time when used as a text field in the DHTLM processing.

I did also look at modifing the DHTML process to define the fields as a date field but that would take at least a half to full day to implement it correctly.

James