Bug #1258 - Error performing currency search.
The web pages have allowed bad currency values in the adoc_currencyField table. I've updated the holdermaint.asp and docmaint.asp pages to enforce currency and date flex fields properly but only when the user edits the field. Invalid currency and date fields will not be flag if the user does not edit them.