Bug #1252 - document tabs not showing
This issues was due to a combination of the Document Type Rules (DTR) and bad currency data. The DTR as executed tries to evaluate conditions users have set up to display or not the document tabs. Because some invalid currency data has been allowed into the system the evaluation failed causing the page to stop rendering at that point. I've fixed the evaluation to ignore invalid currency data.