1410 | Customer Page - Add an <Edit View Preference> to allow the Loan List to Display Automatically Version 14
Many changes to enable the changes requested in ticket.
In ver. 5.3 and older we used to display the list of new loans automatically on the Customer page.  This functionality was put behind a link in ver. 13 to prevent page load wait times.  Some customers would like to see this list without having to click the link.  
We would like to have an option added to the Customer Page's <Edit View Preferences> that will allow this list to automatically display as it did in ver. 5.3.  This option will be defaulted to OFF.