1311 | Editing a Credit Covenant creates a Foreign Key Conflict Version 14
When adding a "Current Value" for a Covenant flexfield it will look to see if a covenant record exists. The software did not recognize correctly when a covenantId was not passed and there for did not create a new covenant record when needed causing the error. I have prevented this misread in the code.