1246 | Duplicate Scanner Names in Scan History Name Dropdown Version 13
Originally a session variable was assigned for userName in the commonlogin.inc file when logging in to AccuAccount and this was then used throughout the system when creating or modifying Document History.  I added a function when the session variable is created, so that the business rule for userName format is adhered too.