1247 | Audits Show Document View Icon Even When De-Selected Version 13
Audits create static HTML pages from the data in the database.  When loaded in the browser that are stored in the cache. If you go and make a change to the audit and then go to look at them again, the files names are the same so the browser loads the page from cache, which contains the old data.  I have made the links dynamic for the page so that the browser is forced to refresh the page and thus show the new data.