1321 | Changing covenants do not obey on page but do under edit customer Version 14
Originally the new value being submitted was checked to see if it was numeric, and if it wasn't it would then use the previous value, however this check was also being used for a blank value.  So I added an additional check to ensure something was in the variable before checking to see if it was numeric, resolving the bug.