Backlog #1375 - Purge Enhancements
Added new admin menu selection for Purging a System. The project now will call an ASPX page to asynchronously call the AccuDoc Purge Library to perform the work and return back to the web page for the user to continue their work. A two step confirmation is presented to the user to prevent accidental purging of the system elements.