Bug - Fix for upload to non-standard folder.
There was an issue with building the paths with non-standard folders (e.g. upload not in the web directory). I've corrected the issue so the paths are built correctly. Also I've added some bulletproofing to check for the existence of the directory and default it back to the standard directory (upload/vol_00001/systemName) to avoid issues with users passing in bad information from their stored cookies.