Bug 660 - [v13.1] Active Task dash panel not filtering by branch security and bank
Modified the program "dashpanelactivetasks.asp" to only allow the viewing of data if the user has been granted access to the exact same bank branch where the data is to be viewed.

I added the include file: include/getUserBranchSecurityArrayMultiBank.inc and used the branch security array that it creates.

I also added a field in the SELECT statement "customerBranchId" and this is the value that I use to test the match of the branch to know if the data matches in the "userbranchsecurity" table.

James