Bug - 1318 - Create New Booked Loan Blanks fields out after missing loan number
I modified 3 programs

a) include/loanmaintdetails.asp
b) customerscan.asp
c) customerscanupdate.asp

In the "customerscan.asp" I added a field named "saveDataOnError" on Line 35.  This is to keep the status of it there was an error that the "include/msghandler.inc" would handle.  In this way I can know if the variables should remain.

In the "customerscanupdate.asp" program I added Session variales that were not already included for the program "include/loanmaintdetails.asp" so that if there was an error the variables that were origionally entered would be returned.

In the program "include/loanmaintdetails.asp" I added the condition to check if the variable "saveDataOnError" was true it desplayed the previously entered variables as the values.

James