1346 | Collateral Flex Field Not Searchable Version 13
Found an copy and paste error that related to the hasSearchableFields boolean that was drawing the count from the accounts and not the collaterals.  During the ticket I found another bug related to the display of the flexfields in the admin area whereby it was showing a flex field as searchable or displayable when they were not.  So I added a change to fielddeflist_collateralbody.inc and fielddeflist_mainbody.inc to fix this.