Bug #1359 - Upload page error at First Savings Bank
I've refactored the code in managing the upload path for consistency between ASP and ASP.Net. I've also added bulletproofing to both ASP and ASP.Net pages to check the user's current upload path to make sure it is valid. If it is not it will return the user to the default upload folder which will be the System Module the user is logged into.