Bug - Fixed issue with the Document Selection and the User defined Holder Flex Field.
There was a copy and paste error introduced from the Holder SQL statement when it was partly copied to the Document select query. The user defined flex field was joining on the Holder's defined flex field and not the DocumentType's defined flex field. This has been corrected.

Project rebuild after library was fixed.