1286 | Print Single Barcode Generates Nothing (left hand edit document tab) Version 13
The page was never set up for waived documents so it has a constraint in the SQL for only active documents. I need to have this constraint when extracting all barcodes, but not on a single barcode.  So I added a parameter to the page this states if the page is for "multi" barcodes the SQL appends the constraints to the SQL statement.  This will alleviate the bug.