1200 | View Unmapped Documents Missing tabs to move docs to Version 14
Changed INNER JOIN to LEFT OUTER on the documenttab as this may not be present. Also added all coding necessary to create a new documenttab and it also will check to see if a document without a physical file already exists and will update it, or it will create a new document instead.