1290 | Unmapped Documents Not Showing Document Tabs to Upload To Version 13
Changed INNER JOIN to LEFT OUTER so the list now populates, as the document tab may not be present. During the move, I added all coding necessary to create a new document tab(if missing) and it also will check to see if a document without a physical file already exists and will update it, or it will create a new document instead.
 
Added an additional script that deletes the redundant document when a new documentId is created, this ensures that the document does not remain in the unmapped document list.