Bug 1318 - Create New Booked Loan Blanks fields out after missing loan number
I only modified two programs - "include/loanmaintdetails.asp" and "customerscanupdate.asp". 

This time I used the javascript to validate if the loan number is entered. If the loan number is not entered then a small window will appear informing the user that the "Loan Number" must be entered. If they have made changes to any of the fields all the changes will remain.

In the program "customerscanupdate.asp" I cleaned up the code to remove the check if the "Loan Number" was blank.

James