Bug 1414 - Loan, Deposit, Trust - When adding a flex field its not following its rules
The error occured when a new loan, deposit or trust was created.  I make the fix to the program "loanmaintdetails.asp" in the area of lines 582 with a the following comment "Get the accurate status of the flex field for when value is NULL"

Now the default status is always accurate matching 100% of what "Flex Field Maintenance" says. 

When a new load, deposit or trust was created there were no matching records in the load table to when the screen was returned for updating so it always made the value off.  Now when this occurs the system reads the specific Flex Field and get the accutate status.

James