Bug #1454 - Upload issue for CBC
The issue was related to bad verbiage indicating the user did not have permission to move files when it should have been related to deleting files from the upload folder. I did refactor the code to cleanup and normalize the upload paths for consistency so this needs to be tested outside of just cleaning up the verbiage.