1458 | Creating New Application causes error | Ongoing | Version 14
Blank characters were not being trimmed from the original value of targetApproverId received in the form.  The form returns "--null, ", and this was being cut to "null, " which was then causing an error because it should say "null".  I added the extra trimming necessary so that the value is returned correctly.