BugBug - 1391 - Inactive Document Barcodes Printing from the Customer Page
Work with the program "HTMLBarcodes.asp" to ensure that all three queries 

a)  creditQuery
b)  loanQuery
c)  collQuery

to make sure that each query has the following columns to know if the group is active or not.  The two columns are

a) da.activationStatus
b) dd.defaultActivationStatus

Last I added an "OUTTER JOIN" to the "collQuery" as the last part of the code:
   LEFT OUTER JOIN documentActivation AS da 
       ON da.loanId=cl.loanId" & _
	AND da.documentTypeId=dd.documentTypeId

Added code the uses the variable "currentActivation" & "printBarCode" to know if the groups is active or not that also deals with NULL values.

James