1428 | Converting Application to Booked Loan Flex field contents drop out Version 13
Added a new sub routine that deletes the existing flexfield for the destination loan, and then it updates the flexfields from the source to destination.  Added this into the merge process for the loans and the collaterals.

1482 | Converting Application to Booked loan - Status choice not holding
Version 13

Added an additional script to ensure that the loan status is updated during the conversion.

838 | Employee Documents - Need to hide documents that are tied based on "additional guarantors/signors"
Version 13

Added to the SQL that extracts data from the primaryBorrowerQuery so that it returns the "employee" field value.  It then uses this variable to establish if the user has permission to view the files in question for the primary borrower.

1344 | View Scan History Sort By Remove Asterisk in drop down
Version 13

Removed the offending asterisk as noted in the ticket.
Cleaned up HTML