Bug 1431 - Error when editing barcode 10 characters or longer (field is 12 characters length)
I modified the program "regmaintedit.asp" to only allow a max of 10 characters on the data entry screen. I left the width as 12 so that there would be some space on each side of the number if 10 digits are entered.

I made the max of 10 characters because the field in the database is defined as an INT field.

James