BugBug - 1391 - Inactive Document Barcodes Printing from the Customer Page
I have make the final code update and correction to the program "HTMLBarcodes.asp".  I made the code more streamlined while fixing the coding error.  I added the following:

        "	ISNULL(da.activationStatus, dd.defaultActivationStatus) AS activationStatus," & _

to each of the three SQL statements and removed an if condition by doing so.

The last error was that I was using the "documentStatus" instead of using the "documentStatusType".  

James