Bug 1331 - Inactivating a User Listed as a Loan Delegate, Lender, etc.
I modified the program "include/loanmaintapplication.asp" to allow the display of the "Loan Delegate" no matter if it is inactive or active but it attached to the loan.

While looking at the subtickets 1381 & 1404 I decided to do a full system search to see where the following code "u.inactive LIKE 'N' AND " was used in conjunction with the "[user]" table.

I found that in the same program "include/loanmaintapplication.asp" the condition was used for "Lender". So if the lender was made "inactive" the program would also fail. So, I modifed the code for this area also.

James