Bug - CBC - Sort on View Account Listing
I updated the program "accountlist.asp" to ensure that the sorting is 100% correct for the "Origination Date". I made sure that the month and day fields were padded with a zero to keep the formating of the date consistant.

In working with getting the date to sort I had to create a custom function to allow the "Loan Number" field to sort correctly. In fixing the date the "Loan Number" field did not want to work any longer. Now, it does and returns the correct results.

James