Bug 1135 - CBC - Sort on View Account Listing
I fixed the sort of the "Loan Number" and other fields by using the custom sort routine written by Paul and used in the program "sh_searchresults.asp"

In my testing I see that the column "Docs?" I find that it does not sort correctly because dhtmlx seems not to sort an icons / image well.  I had a look a the program "sh_searchresults.asp" and found that the same applies to it that the sort is not working for icons / images in the same column.  What I did as not to take any more time is to change the type of field that is is "na" so sorting is not permitted.  In this way 100% of the sortable fields sort correctly.

James