Bug 1509 - Expiration Dates Not Showing on the Audit
I have modified the program "auditbuild.asp" to ensure that the correct expiration date is used. When I was testing the "expDate" from the query " ISNULL(d.expdate, dd.defaultexpdate) AS expDate," I discovered that the value being returned by the query was not match the values on the "expDate" being displayed on the customer page as being returned by the query in the progam "auditbuild.asp".

I worked with the two (2) queries and saw that there was a difference in the retreival of the "expDate". I made the change in the "auditbuild.asp" progam to match the query in the customer page only for the "expDate" and the program "auditbuild.asp" started to return the correct results.

So, I made the same change in each of the three (3) queries in the "auditbuild.asp" program.

The other thing that I discovered was the the results being provided in the ticket did not match companies. I feel to do a proper test would require creating a new audit and match the results.

James