1508 | Unable to Type to Jump Down to a Lender Name by Typing Within Dropdown Field Version 13
Enabled Filtering Mode and Autocomplete on the DHTMLX ComboBox so that you can jump down the list without having to manually scroll down. Just select all the text in the combobox and then begin to type to jump down the list. Found the issue also on loanmaintappadmin.asp and loanmaintapplication.asp.