1501 | AccuAccount Endless Loop/CPU Spike on documentmaintdone.asp Version 13
Added a referer check to see if this page was requested by another page.  If it has no referer then the user is automatically redirected to the dashboard.asp page.  I also added a check to ensure this page is opened within a modal window.  If it was, then the parent page refreshes, closing the modal. If it wasn't then the user is redirect to the dashboard.asp page.