Bug 1504 - After App is Coverted to Booked Loan, Folder Name Shows as the Documents' Title
I have modified the program "convertapplicationupdate.asp" to use "dd.documentSubTypeName" & "dd" stands for "documentDefinitions" table value instead of the "dd.defaultName" value that was in the code before.

Line 512

It took a while to find the error because none of the code that displayed the values was changed for a long time so I had to trace back to see where the value was being writting into the "document" table.

James