Bug 1504 - After App is Coverted to Booked Loan, Folder Name Shows as the Documents' Title
The error that made the initial ticket fail occured in the subroutine"RefactorDocumentHistory".  It was placing a comma after the variable in the UPDATE statement where was only one column being updated.

There is an if condition that would add another column to the UPDATE so I put the comma in that area.

Line 921

Jamess