Bug 1547 - Error When Deleting an Application Status at ANB Bank
, I modified the program "accountstatuslist.asp" to also include this table "actionApplicationStatusFilter" in the conditions so that  if there is a record in the table "actionApplicationStatusFilter" and not in the "loan" table it would not display the delete button.


Line 242 

(SELECT COUNT(*)
    FROM dbo.loan AS l RIGHT OUTER JOIN
               dbo.actionApplicationStatusFilter AS aasf ON l.loanStatusId = aasf.loanStatusId"       
    WHERE (aasf.loanStatusId = ls.StatusId) )AS accountCount

A feature that would be good to add to the delete process is a "Are You Sure" window asking the user if they are sure that they want to delete the record.  Right now if you press on the delete button it will just remove the record.  

James