Bug - 1416 - Export Error at Valley Business Bank
I modified the program "exportsearch.asp" to add the condition that if there are Records that has "Is Searachable" selected this message would be displayed and prevent errors before the export is being created:

"There are no searchable Record Names defined. Please go to "Record Definition Maintenance" and assign the desired data as searchable."

I added the variable "processedLoop" as a true / false condition.

James