1576 | Development hotfix for 2014 menu for Woodforest National Bank. Version 14
Rewrote the menu to user older style html tags and javascript, to ensure that the menu works correctly in IE5 document mode.  Had to remove secondary sub menu for "Notice Letters" as the technology would not work correctly for IE5. Added the "Notice Letter" secondary sub menu elements to the "Exceptions" Menu instead.