Bug 1544 - Unable to See Document Definitions In AccuDoc at BOBV
I have made it very difficult for non-compliant characters to be saved in the "documentTypeName" ie "Unique Type Name (optional)" field to always remain compliant even if the user types data into the field.

The end user will not receive any mesages regarding the processing because it is done on the backend of the system just before the data is actually stored.

I have modified two programs "doctypeupdate.asp (line112)" and "include/xmlexport.inc (line 207)" with the same code to ensure that the compliance is exactly the same.

James