Bug - 1561 - Collateral Status Changes when you Edit a Loan
I updated the program "customerscanupdate.asp" to not update the status of the loan is updated.  So, now when a loan is updated the callaterals status will remain the same.

I removed this line of code "collateralRS("loanStatusid")        = loanRS("loanStatusid")" which started on line 601 of the program.

When I spoke with Mike he thought that there might be a stored procedure also in this process but the only stored procedure that I found was "spAddNewAccountDocuments"  but there in no reference to UPDATING only INSERTING in the stored procedure.

James