Bug 1573 - Changing a Loan Number Causing Scanned Image Folder to be Deleted
This fix will require that the end to resolve the duplicate folder name in the "scanned_images" folder before they will be allowed to change the loan number. When they try to change the loan number they will receive the following error message:

A scanned image folder: [ destFolder ] already exists for the new Loan no: [xxxxxx] has [xx] sub folders and has [xx] folders."

The text that is brackets will be replaced with the actual values / counts from the system. This fix is really so that the system does not by default remove the folders when there is content in them like it did by default.

I modified two programs (1) /include/common.inc (2) changeloannumberupdate.asp.

In the (1) I changed the REPLACE statement in the function [replaceSpecialCharacters] for replacing the space with a "-" not an underscore as it was before. This is because the last line of the function does a REPLACE for an underscore with nothing... 

In the (2) I added the code to check to see if the folder exists and if it does get the details of the folder so that they can be displayed to the user in an error message. [line 106 - 159]

James