Bug 1544 - Unable to See Document Definitions In AccuDoc at BOBV
I have added code to not allow the ASP reserved word "OPEN" to be used as the first word in the field. If it is used it will place the word "flex_" in the option field name in front of the work "OPEN". (Line 126)

Also, I made it so that the last character of the optional field is not an underscore. (Line 137)

I have made it very difficult for non-compliant characters to be saved in the "documentTypeName" ie "Unique Type Name (optional)" field to always remain compliant even if the user types data into the field.

The end user will not receive any mesages regarding the processing because it is done on the backend of the system just before the data is actually stored.

I have modified two programs "doctypeupdate.asp (line112)" and "include/xmlexport.inc (line 207)" with the same code to ensure that the compliance is exactly the same.

James