Bug 1416 - Export Error at Valley Business Bank
I updated the program "exportsearch.asp" to ensure that the titles / groups are correctly grouped and the the message if there are not any searchable fields to be displayed per group. Also, I have written code the will display the search button only if there are search fields. The cancel and search button appear at the top and the bottom of the list. The new code does a precheck to know if to display the search button at the top.

Line 87 - 125

James