Bug 1568 - AccuApproval - When no default statuses are selected in admin, and create new app it still brings in a default analysis status
I have made the change in the system so that if there is not an "Analysis Status" default is selected the first value in the list box on the "App Admin" tab on the edit screen will have the text "No Default Value Selected". On the customer page in the space for the value of "Analysis Status" it will show "---".

This process will enter a NULL value for the "Analysis Status" until it is selected.

Regarding "Approval Status" the system will not do a check to see if a default has been asssigned. The list box will always display the first value based on an alphabetical sort.

I have modified two program (1) include/loanmaintappadmin.asp [line 128 - 169] (2) include/cust_accountheaderloanapp.inc [Line 71]

(1) This program "include/loanmaintappadmin.asp" does a read of the database to see if there are any defaults assigned

(2) This program "include/loanmaintappadmin.asp" does a check to see if the value is NULL and it so it will display the "---" or the value from the database.